Fix primary guardians controller edit to handle empty child case properly

+++ OFRPDMS/OFRPDMS/Controllers/PrimaryGuardiansController.cs
@@ -181,8 +181,11 @@ namespace OFRPDMS.Controllers
                     if (primaryguardian.Children[i].Delete == true || (primaryguardian.Children[i].Birthdate == null && primaryguardian.Children[i].FirstName == null
                         && primaryguardian.Children[i].LastName == null && primaryguardian.Children[i].RelationshipToGuardian == null))
                     {
-                        Child child = context.Children.Find(primaryguardian.Children[i].Id);
-                        context.Children.Remove(child);
+                        if (primaryguardian.Children[i].Id != 0)
+                        {
+                            Child child = context.Children.Find(primaryguardian.Children[i].Id);
+                            context.Children.Remove(child);
+                        }
                         primaryguardian.Children.RemoveAt(i);
                     }
                     else
